% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTools_fr.R
\name{cleantext_fr}
\alias{cleantext_fr}
\title{Clean Text (French)}
\usage{
cleantext_fr(text, language = "french", stop.words = TRUE)
}
\arguments{
\item{text}{character text to be cleaned}

\item{language}{string. Default "french".}

\item{stop.words}{logical. Default TRUE}
}
\value{
a character vector
}
\description{
Basic text cleaning for French
}
\keyword{internal}
