% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTools_fr.R
\name{ctxpand_fr}
\alias{ctxpand_fr}
\title{French Contraction Expander}
\usage{
ctxpand_fr(text)
}
\arguments{
\item{text}{a character vector of texts.}
}
\value{
a character vector
}
\description{
Expands French Contractions
}
\keyword{internal}
