% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictTools_fr.R
\name{french_dictionaries}
\alias{french_dictionaries}
\title{French Dictionaries for Politeness Features}
\description{
French-language dictionaries for politeness feature extraction.
These dictionaries are loaded as internal package data via sysdata.rda.
}
\keyword{internal}
