% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTokenSets_fr.R
\name{getTokenSets_fr}
\alias{getTokenSets_fr}
\title{Extracting Tokens from Natural Language (French)}
\usage{
getTokenSets_fr(text, parser = c("none", "spacy"), num_mc_cores = 1)
}
\arguments{
\item{text}{a character vector of texts.}

\item{parser}{character Name of dependency parser to use.}

\item{num_mc_cores}{integer Number of cores for parallelization. Default is 1.}
}
\value{
list of compiled POS-tagged items.
}
\description{
Return tokens (words or POS tags) from French natural language.
}
\keyword{internal}
