% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacyParser_fr.R
\name{spacyParser_fr}
\alias{spacyParser_fr}
\title{Spacy Parser (French)}
\usage{
spacyParser_fr(txt)
}
\arguments{
\item{txt}{a character vector of texts.}
}
\value{
list of compiled POS-tagged items.
}
\description{
Return POS tags from French natural language.
}
\keyword{internal}
